Organised Statement builder interfaces into a subpackage
