+++ server/src/graql/query/pattern/statement/Statement.java
@@ -25,6 +25,11 @@ import grakn.core.graql.query.pattern.Pattern;
 import grakn.core.graql.query.pattern.property.RelationProperty;
 import grakn.core.graql.query.pattern.property.TypeProperty;
 import grakn.core.graql.query.pattern.property.VarProperty;
+import grakn.core.graql.query.pattern.statement.builder.StatementAttributeBuilder;
+import grakn.core.graql.query.pattern.statement.builder.StatementInstanceBuilder;
+import grakn.core.graql.query.pattern.statement.builder.StatementRelationBuilder;
+import grakn.core.graql.query.pattern.statement.builder.StatementThingBuilder;
+import grakn.core.graql.query.pattern.statement.builder.StatementTypeBuilder;
 import graql.exception.GraqlException;
 import graql.util.StringUtil;
 import org.slf4j.Logger;
