+++ server/src/graql/query/pattern/statement/builder/StatementTypeBuilder.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.core.graql.query.pattern.statement;
+package grakn.core.graql.query.pattern.statement.builder;
 
 import grakn.core.graql.query.Graql;
 import grakn.core.graql.query.Query;
@@ -32,6 +32,8 @@ import grakn.core.graql.query.pattern.property.ThenProperty;
 import grakn.core.graql.query.pattern.property.TypeProperty;
 import grakn.core.graql.query.pattern.property.VarProperty;
 import grakn.core.graql.query.pattern.property.WhenProperty;
+import grakn.core.graql.query.pattern.statement.Statement;
+import grakn.core.graql.query.pattern.statement.StatementType;
 
 import javax.annotation.CheckReturnValue;
 import javax.annotation.Nullable;
@@ -39,7 +41,7 @@ import javax.annotation.Nullable;
 /**
  * Type Statement Properties
  */
-interface StatementTypeBuilder {
+public interface StatementTypeBuilder {
 
     /**
      * @param name a string that this variable's label must match
